+++ src/acconeer/exptool/a121/_core/peripherals/communication/exploration_protocol/__init__.py
@@ -3,5 +3,6 @@
 
 from ._factory import get_exploration_protocol
 from ._latest import ExplorationProtocol, ExplorationProtocolError
+from ._no_5_2_mhz import ExplorationProtocol_No_5_2MHz_PRF
 from ._no_calibration_reuse import ExplorationProtocol_NoCalibrationReuse
 from .server_error import ServerError